Remove cleanup and skip from todo
I really don't wanna mess with those until there is a good argument, at
the moment it feels beyond the scope of eventually. If users want to use
cleanup and/or skip they will still have access to the parent t and can
work around this with deferring, returning early, etc.